переписал компонент InputField под новые требования
